api/requirements.txt
Flask==3.1.2
mlflow==2.18.0
protobuf<6
numpy
pandas
scikit-learn==1.5.2
lightgbm==4.3.0
pytest
python-dotenv==1.2.1
gunicorn